+++ resolwe/elastic/tests/test_app/migrations/0001_initial.py
@@ -3,6 +3,7 @@
 from __future__ import unicode_literals
 
 from django.db import migrations, models
+import django.db.models.deletion
 
 
 class Migration(migrations.Migration):
@@ -40,6 +41,7 @@ class Migration(migrations.Migration):
                 ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('name', models.CharField(max_length=30)),
                 ('dependencies', models.ManyToManyField(to='test_app.TestDependency')),
+                ('dependency', models.ForeignKey(null=True, on_delete=django.db.models.deletion.SET_NULL, related_name='main_dep', to='test_app.TestDependency')),
             ],
         ),
         migrations.CreateModel(