+++ docs/CHANGELOG.rst
@@ -6,6 +6,15 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+Fixed
+-----
+- Pin ``django-priority-batch`` to version ``1.1`` to fix compatibility issues
+
+
 ===================
 16.0.0 - 2019-04-16
 ===================