+++ docs/CHANGELOG.rst
@@ -20,6 +20,7 @@ Changed
 Added
 -----
 - Support Python 3.7
+- Support forward and reverse many-to-one relations in Elasticsearch
 
 
 ===================